feat: Add support for sending mint messages
This commit adds support for sending mint messages in the Discord bot. It checks the value of the `SEND_MINT_MESSAGES` environment variable and only sends the mint message if it is set to "true". This feature will allow the bot to send notifications when the Peercoin Foundation mints new Peercoin.