chore: Update dependencies and fix Discord bot message sending
This commit updates the dev dependencies to the latest versions, specifically the lints package to version 4.0.0 and the test package to version 1.16.0. Additionally, the dependencies hive and nyxx are updated to versions 2.2.1 and 6.2.1 respectively.

In the lib/discord.dart file, the code is modified to use the NyxxGateway client instead of the deprecated NyxxWebsocket bot. The client is now created using the Nyxx.connectGateway() method and the necessary plugins are registered. The sendMintMessage() function is also updated to use the sendMessage() function asynchronously.

These changes ensure that the dependencies are up to date and fix the issue with sending messages in the Discord bot.